get optimal feedrate
